remove encode/decode funcs
