+++ app/plexus-agent/cmd/loglevel.go
@@ -16,6 +16,7 @@ limitations under the License.
 package cmd
 
 import (
+	"encoding/json"
 	"fmt"
 	"strings"
 
@@ -41,7 +42,7 @@ debug, info, warn, or error
 		fmt.Println("setting daemon log level to", args[0])
 		ec, err := agent.ConnectToAgentBroker()
 		cobra.CheckErr(err)
-		data, err := agent.Encode(plexus.LevelRequest{Level: strings.ToLower(args[0])})
+		data, err := json.Marshal(plexus.LevelRequest{Level: strings.ToLower(args[0])})
 		cobra.CheckErr(err)
 		cobra.CheckErr(ec.Publish(agent.Agent+plexus.LogLevel, data))
 		cobra.CheckErr(ec.Flush())
