+++ internal/agent/publish.go
@@ -1,11 +1,13 @@
 package agent
 
 import (
+	"encoding/json"
 	"log"
 	"log/slog"
 
 	"github.com/devilcove/boltdb"
 	"github.com/devilcove/plexus"
+	"github.com/nats-io/nats.go"
 )
 
 func publishDeviceUpdate(self *Device) {
@@ -15,7 +17,7 @@ func publishDeviceUpdate(self *Device) {
 		slog.Error("not connected to server")
 		return
 	}
-	data, err := Encode(plexus.Peer{
+	data, err := json.Marshal(plexus.Peer{
 		WGPublicKey:   self.WGPublicKey,
 		PubNkey:       self.PubNkey,
 		Version:       self.Version,
@@ -41,7 +43,7 @@ func publishListenPortUpdate(self *Device, network *Network) {
 		slog.Error("not connected to server")
 		return
 	}
-	data, err := Encode(plexus.ListenPortResponse{
+	data, err := json.Marshal(plexus.ListenPortResponse{
 		ListenPort:       network.ListenPort,
 		PublicListenPort: network.PublicListenPort,
 	})
@@ -61,7 +63,7 @@ func publishNetworkPeerUpdate(self Device, peer *plexus.NetworkPeer) error {
 	if serverEC == nil {
 		return ErrNotConnected
 	}
-	data, err := Encode(peer)
+	data, err := json.Marshal(peer)
 	if err != nil {
 		return err
 	}
@@ -125,4 +127,30 @@ func checkin() {
 		return
 	}
 	log.Println("checkin response from server", serverResponse.Message)
+}
+
+func publishErrorMessage(conn *nats.Conn, subj string, err error) {
+	response := &plexus.MessageResponse{
+		Message: "error" + err.Error(),
+	}
+	bytes, err := json.Marshal(response)
+	if err != nil {
+		slog.Error("invalid message respone", "error", err, "data", response)
+	}
+	if err := conn.Publish(subj, bytes); err != nil {
+		slog.Error("publish error", "error", err)
+	}
+}
+
+func publishMessage(conn *nats.Conn, subj string, msg string) {
+	response := &plexus.MessageResponse{
+		Message: msg,
+	}
+	bytes, err := json.Marshal(response)
+	if err != nil {
+		slog.Error("invalid message response", "error", err, "data", response)
+	}
+	if err := conn.Publish(subj, bytes); err != nil {
+		slog.Error("publish message", "error", err)
+	}
 }
